[Fix build] Revert EncTool and Remove unused ``CheckIntegrity``
The CheckIntegrity code was supposed to be used within Game Conversion feature, but the functionality has been completely replaced by HonkaiRepair.
This code also causing the build to fail due to uncomplete code/changes that happened in EncTool.

This code also has no references to any component to the launcher so this code is completely unused.

This commit also revert the EncTool submodule to the main repository with some adjustment being done to the core codes to fix build issues