Lower Collapse process to BelowNormal when game is running
This commit add a toggle (default off) to lower Collapse process to Below Normal when game is launched, this is beneficial for lower-end PC user.

When this feature is enabled and no game is running, also set Priority Boost to enabled, allowing Windows to bump Collapse priority when its in focus.