Implement GameBoost
This feature will boost Game's process priority to Above Normal, in hope it gets better resource by Windows scheduler and gives better in-game performance.

Why not high priority you may ask? High priority is found to be quite flaky if being put in a game process, mainly crashing but can also cause system to freeze due to all the resource is being taken by the game.

Also as a note, it waits for 20 seconds before applying it due to HI3 seems to spawn another process as a main process instead of using the process it got called as one (whyyyy)