Making Background Image loading fully asynchronous
This also move the download call outside of API's fetching mechanism to ensure the API will be called first and finish, making the token not cancelled due to timeout of loading the background image.

I also make the timeout duration to be separated from API loading. It will be using 60 seconds for the timeout instead of 10 seconds for API loading.

This change also change the size limit for each cached event panel images and background image for logic comparison.
The launcher will download the event panel images and background image if the file doesn't exist and the size is less than 4 KB and 1 MB respectively