[DRPC] Fix odd behavior with play and idle activity
There is an old workaround that was forgotten to be deleted to handle the SetActivity call when it was called in a fast loop. As now we dont use fast loop, the call were getting returned early and RPC is not modified.

Also add force idle after the game is stopped