Prevent Tray Notification to halt the app
This usually happens at launch where Tray is not yet fully initialized (Windows, deal with it) and the ShowNotification function throw an error because of it that would halt the app from starting. This commit catch the exception and completely throw it out of the windows (pun intended, frick you windows)

Is this commit message too long? maybe. maybe not. maybe im just yapping. who knows. i know, i guess.