Fix Zenless not showing log in console
Apparently the log file path has changed and its a dynamically assigned name based on when the function in game is called.

This change will wait until the file generated in the directory, make sure its the correct file by checking it in the pattern, then proceeds to read the file.

This commit also fixes the FileSharingException issue due to Collapse trying to get a Write access to the file, as we only do a read operation, we will ask ONLY for read access and let the FileSharing as ReadWrite to allow the game to write onto the log.
Garbled data read from old log is prevented by simply moving the old log file to let the log to be read as fresh as possible, "it's mint bud".

I definitely not overcomplicating this issue fix, I maybe a tini-tiny bit, maybe more. Alright I admit I overcomplicated this by ALOT, happy?