Remove warnings from ThisAssembly and GitInfo This will be long but first and foremost We are very thankful for both packages maintainer for providing us the packages that we can use in our projects.
With that got out of the way, WHY in the world would you use warnings for those? SURELY there will be a better way to do this. I do fully understand that maintaining a huge amount of libraries used by many is hard, and who am I talking for just a novice that tries to code for my own enjoyment while mostly losing money keeping this project alive. I know our IDE of choice is "annoying" for having this warning appears, but if disabling this warning in our project bother any of the maintainers please feel free to contact me (contact infos in my profile) and I'll remove the library from our project.
Thank you