[Imp] Implement new Sophon Patch Mode (#714)
This PR implements a new update mode for Sophon which is currently in
A/B Testing phase for the official HoYoPlay launcher.
Sophon Patch Update mode brings a much improved download pipeline and
update efficiency as this method involves not only Copy-Over method, but
also Patch-Over method via HDiffPatch.

This results a much smaller download footprint compared to the old
Copy-Over only method. As an example of this, we compared Genshin Impact
5.4.0 -> 5.5.0 update between the Old and New method.

The old method requires 15.97 GB of data to download. Meanwhile, the new
method only requires 5.92 GB of data to download, which is **62.9%** of
size reduction over old method.

Note: In this test, we installed two VOs: en-us + ja-jp

![image](https://github.com/user-attachments/assets/67b8e1f7-d33e-4086-91aa-09125a53728c)