Re-use bitmap data on loading bg image accent color
This is achieved by switching BitmapSource for the background image to WriteableBitmap. With this, we can access an immediate buffer on the image instance instead of re-reading it to Bitmap twice.

We also add an extension on Win32.WinRT submodule to cast IBufferByteAccess from WriteableBitmap.PixelBuffer (IBuffer) since it wasn't accessible from WindowsAppSDK out of the box