[EMERGENCY] Disable MobileMode on all games.
Due to the recent ban for using this feature, we are disabling it for now by force.
While the ban seems to be only for co-op, we are not taking it lightly so we disable this feature until we are sure that it is safe to use.

If you have similar issue with this feature and/or other feature that causes something to your account, please contact us immediately.

If you are from HoYo, thank you, for doing this, to your players, I'm sure that game cloud subscription is worth it.